Remove usage of CamelCase on field names
