+++ ui/src/components/shared/field-map.js
@@ -22,16 +22,16 @@ export default {
 	Title,
 	Radio,
 	Checkbox,
-	ImageSelect,
+	Image_Select: ImageSelect,
 	Select,
-	TextArea,
+	Text_Area: TextArea,
 	Link,
 	Hidden,
 	HTML,
 	Image,
-	ImageGallery,
+	Image_Gallery: ImageGallery,
 	Video,
-	PostQuacker,
+	Post_Quacker: PostQuacker,
 	Post_List: PostList,
 	Swatch_Select: SwatchSelect,
 	Group,
