+++ ui/src/data/blueprint-multi.json
@@ -381,7 +381,7 @@
           ]
         },
         {
-          "type": "ImageGallery",
+          "type": "Image_Gallery",
           "label": "Image Gallery",
           "name": "gallery",
           "description": "Gallery description goes here",
